Initial Phase 001 Commit
This is our initial Phase 001 commit; at this point we can create random players, create random teams, and play individual games, keeping track of a handful of individual and team stats. Next on the list is the League and Schedule classes. To complete Phase 001, we have the Season, Playoff, and Offseason classes.